% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-matching.R
\name{matchxrf}
\alias{matchxrf}
\title{Matches a set of XRF peaks to elemental transition lines}
\usage{
matchxrf(xrfpeakdata, sample.elements = "", match.interval = 0.1)
}
\arguments{
\item{xrfpeakdata}{Dataframe containing the observed line energies and
expected to have the following column names and types:
$ peak      : peak number (integer)}

\item{sample.elements}{limit matching to only these elements (vector of strings)
Defaults to empty string (i.e., matching against all elements).}

\item{match.interval}{Matches will only be considered inside this energy window.
Defaults to 0.1 keV, meaning it will match within +/- 0.05 keV.}
}
\value{
Dataframe with the following columns
  $ Z                     numeric, element number
  $ element               string, chemical element
  $ IUPAC.formatted       string, IUPAC label of ref transition line
  $ transition.keV        numeric, reference position of transition line
  $ energydiff            numeric, delta energy between obs and ref
  $ peak                  numeric, the peak number (peak fit)
  $ kernel                numeric, the kernel number (peak fit)
  $ x                     obs position of XRF peak
  $ within.match.interval boolean, indicating whether match is inside interval
  $ unique_id             string, uniquely identifiying each row
  $ best.match            <<<<<
}
\description{
Function for matching a set of XRF peaks to their nearest
matches (on the number line) from atomic transition lines
across the whole periodic table (Z = 4 to Z = 103) or a
subset of the periodic table.
For example for matching an experimentally observed XRF peak
to the nearest-lying literature values.
}
\examples{
\dontrun{matchxrf(xrfpeakdata, sample.elements = c("Zn", "O"))}
}
