% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xraylabels}
\alias{xraylabels}
\title{IUPAC and Siegbahn notation of electronic transitions}
\format{
A data frame with 70 rows and 9 variables (all character class):
\describe{
  \item{IUPAC}{Transition label in the recommended IUPAC notation}
  \item{shell.filled}{Electronic shell where the transitioning electron starts}
  \item{subshell.filled}{Electronic subshell of the same}
  \item{shell.source}{Electronic shell where the transitioning electron ends up}
  \item{subshell.source}{Electronic subshell of the same}
  \item{Siegbahn}{Transition label in Siegbahn notation}
  \item{Siegbahn.short}{Short form Siegbahn notation}
  \item{IUPAC.formatted}{LaTeX-formatted label in IUPAC notation}
  \item{Siegbahn.formatted}{LaTeX-formatted label in Siegbahn notation}
}
}
\source{
\url{https://doi.org/10.6084/m9.figshare.1168939.v1}
}
\description{
A dataframe with IUPAC transition names and their corresponding
  labels in the old Siegbahn notation. Also with LaTeX-formatted strings.
}
\author{
Taha Ahmed
}
